+++ spec/acceptance/aphlict_spec.rb
@@ -131,4 +131,9 @@ RSpec.describe 'phabricator::aphlict' do
   context command('echo | wscat ws://127.0.0.1:22280') do
     its(:exit_status) { is_expected.to be_zero }
   end
+
+  context command('logrotate --debug /etc/logrotate.d/aphlict') do
+    its(:exit_status) { is_expected.to be_zero }
+    its(:stdout) { is_expected.to be_empty }
+  end
 end